big player changes
Heads up: powerups don't need their separate scripts with this setup, you just need to ensure the powerup is named correctly in the scene.

Also, we can probably add the line renderer back to the player if you want - the issue was the camera the whole time -> it's not happy when the player runs into things apparently.